+++ tests/conftest.py
@@ -6,12 +6,11 @@ import pytest
 from tests.plugins.tortoise_orm import cleanup, init_tortoise
 
 
-@pytest.fixture
+@pytest.fixture()
 def template_dir(tmp_path: pathlib.Path) -> pathlib.Path:
     return tmp_path
 
 
-@pytest.mark.asyncio
 @pytest.fixture()
 async def scaffold_tortoise() -> AsyncGenerator:
     await init_tortoise()